+++ zebra-chain/CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   renamed to `Overflow`.
 - `parameters::network::subsidy::num_halvings` was renamed to `halving`.
 - `transparent::Output::new_coinbase` was renamed to `new`.
+- `NoteCommitmentSubtreeIndex` now derives `schemars::JsonSchema` (#10201)
 
 ## [5.0.0] - 2026-02-05