+++ zebra-chain/CHANGELOG.md
@@ -37,7 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `BoundedVec` re-export.
 - `OrchardActions` trait with `actions()` method.
 - `ConfiguredFundingStreamRecipient::new_for()` method.
-- `strum` dependency
+- `strum`, `bounded-vec` dependencies.
 
 ### Changed